Casted away a few compiler warnings - tons too go.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@11117 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/collide/opcode/OPC_FPU.h
@@ -57,7 +57,7 @@
 #include <math.h>
         inline_ float FastSqrt(float square)
         {
-                return sqrt(square);
+                return (float) sqrt(square);
         }
 
 	//! Saturates positive to zero.
